RDMPDEV-1580 Simplified various tests (especially setup/teardown)

+++ Rdmp.Core.Tests/CohortCreation/AggregateFilterPublishingTests.cs
@@ -41,7 +41,8 @@ namespace Rdmp.Core.Tests.CohortCreation
         [TearDown]
         public void DeleteFilter()
         {
-            _container.DeleteInDatabase();//cascades to filter
+            if(_container.Exists())
+                _container.DeleteInDatabase();//cascades to filter
         }
 
         [Test]
